fix 不允许删除工单内的最后一个子工单。以及，新增（修改）纳管库的时候不允许使用已有的ip-port-(sid|service_name)

+++ restful_api/views/offline/sub_ticket.py
@@ -133,6 +133,8 @@ class SubTicketHandler(TicketReq):
         sub_ticket = TicketSubResult.objects(statement_id=statement_id).first()
         if not sub_ticket:
             return self.resp_bad_req(msg=f"找不到子工单编号为{statement_id}")
+        if TicketSubResult.objects(work_list_id=sub_ticket.work_list_id).count() == 1:
+            return self.resp_forbidden(msg="删除失败，工单应该有至少一条语句。")
         sub_ticket.delete()
         self.resp_created(msg="删除成功。")
